Cleanup Examples & Reorg
The chat-service and echo-service have been moved into examples.

Added examples section to toml

Build example services, allowing the test server to just link and go.

It became clear that adding services was hindered by having to mess with the settings. That is now fixed, by changing from enums to a string.