Replace scheduler (#7)
Major change to the scheduler and executor. Select didn't scale well, so it was dropped. Instead there is a tighter coupling between the channel and executor. Dropping of a machine has moved from the executor to the scheduler.

fmt --all
removed usage of select for machine readiness.
re-worked the channel sender to add tasks
bench and test and ad-hoc testing all passing

Prepare for 0.1.2 release
Removed Select Usage for machines
Fmt, Clippy, README, Changes